+++ packages/vue-color-scheme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vue-color-scheme
 
+## 0.15.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vue-utils@0.15.13
+
 ## 0.15.15
 
 ### Patch Changes