+++ packages/vue-scroller/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vue-scroller
 
+## 0.15.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vue-utils@0.15.13
+
 ## 0.15.12
 
 ### Patch Changes