+++ packages/vue-stack/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @fastkit/vue-stack
 
+## 0.17.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vue-utils@0.15.13
+  - @fastkit/vue-body-scroll-lock@0.2.14
+  - @fastkit/vue-click-outside@0.2.13
+  - @fastkit/vue-resize@0.2.13
+  - @fastkit/vue-transitions@0.2.14
+
 ## 0.17.2
 
 ### Patch Changes